Create CopyToOtherPartyHelper.test
